Debug encoder

+++ include/Sample.h
@@ -26,7 +26,7 @@ private:
    
 public:
     Sample(time_t _epoch, uint16_t _depth, uint8_t _frequency);
-    Sample(uint8_t _hour, uint8_t _minute,  uint8_t _day, uint8_t _month, uint16_t _year, uint16_t _depth, uint8_t _frequency);
+    Sample(uint8_t _hour, uint8_t _minute,  uint8_t _day, uint8_t _month, uint16_t _year, int _depth, uint8_t _frequency);
     uint8_t get_day();
     uint8_t get_hour();
     uint8_t get_minute();
@@ -39,8 +39,8 @@ public:
 
 // =============== SampleS manipulation ==================
 
-uint8_t add_sample(time_t _epoch, uint16_t _depth, uint8_t _frequency);
-uint8_t add_sample(uint8_t _hour, uint8_t _minute,  uint8_t _day, uint8_t _month, uint16_t _year, uint16_t _depth, uint8_t _frequency);
+uint8_t add_sample(time_t _epoch, int _depth, uint8_t _frequency);
+uint8_t add_sample(uint8_t _hour, uint8_t _minute,  uint8_t _day, uint8_t _month, uint16_t _year, int _depth, uint8_t _frequency);
 time_t timeToEpoch(uint8_t _hour, uint8_t _minute, uint8_t _day, uint8_t _month, uint16_t _year);
 time_t get_next_sample_time();
 Sample get_sample(uint8_t number);
